backend/requirements.txt
fastapi==0.115.5
email-validator==2.2.0
uvicorn[standard]==0.32.1
supabase==2.10.0
anthropic==0.40.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
httpx==0.27.2
